requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1

# Task Queue
celery==5.3.4
kombu==5.3.4

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
playwright==1.40.0
scrapy==2.11.0
httpx==0.25.2

# ML & Deep Learning
torch==2.1.1
torchvision==0.16.1
tensorflow==2.15.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0
imagehash==4.3.1

# NLP
transformers==4.36.2
sentence-transformers==2.2.2
spacy==3.7.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
nltk==3.8.1

# APK Analysis
androguard==3.4.0a1
pyaxmlparser==0.3.28

# PDF Generation
reportlab==4.0.7
PyPDF2==3.0.1
fpdf2==2.7.6

# Image Processing
imageai==3.0.3
face-recognition==1.3.0

# API & HTTP
aiohttp==3.9.1
python-multipart==0.0.6

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Monitoring & Logging
prometheus-client==0.19.0
loguru==0.7.2

# Utils
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
pyyaml==6.0.1
click==8.1.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
faker==20.1.0

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3

# Google Play Store
google-play-scraper==1.2.4

# App Store
app-store-scraper==0.3.5

# Async
asyncio==3.4.3
aiofiles==23.2.1

# Data Validation
validators==0.22.0

# Graph Analysis
networkx==3.2.1
